requirements.txt
Flask==3.0.3
tencentcloud-sdk-python>=3.1.89
gunicorn>=21.2.0
